Move files related to potion effects to consumable package
