+++ Server/src/main/java/Server/plugin/consumable/effects/SetAttributeEffect.java
@@ -1,8 +1,8 @@
-package plugin.consumable.potion.effects;
+package plugin.consumable.effects;
 
 import core.game.node.entity.player.Player;
 import core.game.world.GameWorld;
-import plugin.consumable.potion.PotionEffect;
+import plugin.consumable.PotionEffect;
 
 public class SetAttributeEffect extends PotionEffect {
     String attrString;
